feat(AccountModal): add purchase history menu item
